Updated Project Layout
