+++ studystore/FindingFiveStudyStoreUser.py
@@ -1,9 +1,10 @@
 class FindingFiveStudyStoreUser:
-    def __init__(self, userID, numCredits, ownedStudies, viewedStudies):
+    def __init__(self, userID, numCredits, ownedStudies, viewedStudies, wishList):
         self.userID = userID
         self.numCredits = numCredits
         self.ownedStudies = ownedStudies  # list of FFSS studies
         self.viewedStudies = viewedStudies  # list of FFSS studies
+        self.wishList = wishList  # wish list of FFSS studies
 
     def get_userId(self):
         return self.userID
@@ -17,6 +18,9 @@ class FindingFiveStudyStoreUser:
     def get_viewedStudies(self):
         return self.viewedStudies
 
+    def get_wishList(self):
+        return self.wishList
+
     def set_userId(self, newUserID):
         self.userID = newUserID
 
@@ -25,6 +29,9 @@ class FindingFiveStudyStoreUser:
 
     def set_ownedStudies(self, newOwnedStudies):
         self.ownedStudies = newOwnedStudies
+
+    def set_wishList(self, newWishList):
+        self.wishList = newWishList
 
     def set_viewedStudies(self, newViewedStudies):
         self.viewedStudies = newViewedStudies
